+++ packages/utils/src/index.ts
@@ -1,12 +1,13 @@
 import debug from 'debug'
 import fs from 'fs-extra'
-import { globby } from 'globby'
 import hash from 'hash-sum'
 import ora from 'ora'
 import colors from 'picocolors'
+import picomatch from 'picomatch'
+import * as tinyglobby from 'tinyglobby'
 import path from 'upath'
 
-export { debug, colors, fs, globby, hash, ora, path }
+export { debug, colors, fs, hash, ora, path, picomatch, tinyglobby }
 
 export * from './console/index.js'
 export * from './module/index.js'