+++ CHANGELOG.md
@@ -1,3 +1,60 @@
+# [2.0.0-rc.26](https://github.com/vuepress/core/compare/v2.0.0-rc.25...v2.0.0-rc.26) (2025-10-14)
+
+
+### Bug Fixes
+
+* **cli:** fix dev pages watcher ([#1662](https://github.com/vuepress/core/issues/1662)) ([a3ad908](https://github.com/vuepress/core/commit/a3ad90878f725e9654045eb759f44001c7fdabf2))
+
+
+
+# [2.0.0-rc.25](https://github.com/vuepress/core/compare/v2.0.0-rc.24...v2.0.0-rc.25) (2025-08-26)
+
+
+### Code Refactoring
+
+* **utils:** migrate from globby to tinyglobby ([#1660](https://github.com/vuepress/core/issues/1660)) ([5eea315](https://github.com/vuepress/core/commit/5eea31546e5442f2de0a3eb3872a9208f505106e))
+
+
+### Features
+
+* **cli:** upgrade chokidar to v4 (close [#1605](https://github.com/vuepress/core/issues/1605)) ([#1661](https://github.com/vuepress/core/issues/1661)) ([c2e1521](https://github.com/vuepress/core/commit/c2e15217f86a6c84f5774f4a4420c658d4961b21))
+
+
+### BREAKING CHANGES
+
+* **utils:** the `globby` export has been removed, please use `tinyglobby` instead
+
+
+
+# [2.0.0-rc.24](https://github.com/vuepress/core/compare/v2.0.0-rc.23...v2.0.0-rc.24) (2025-06-30)
+
+
+### Build System
+
+* drop node 18 ([#1655](https://github.com/vuepress/core/issues/1655)) ([afad5d0](https://github.com/vuepress/core/commit/afad5d0cf5bd8e2b978e8d4d6d53a718578aacaa))
+
+
+### Features
+
+* **bundler-vite:** upgrade vite to v7 ([#1656](https://github.com/vuepress/core/issues/1656)) ([d4e58c3](https://github.com/vuepress/core/commit/d4e58c3c28da42b7990e22d6dad540d2484e550c))
+
+
+### BREAKING CHANGES
+
+* drop node 18
+* **utils**: `TEMPLATE_RENDERER_OUTLETS` has been replaced with `TemplateRendererOutlet`
+
+
+
+# [2.0.0-rc.23](https://github.com/vuepress/core/compare/v2.0.0-rc.22...v2.0.0-rc.23) (2025-05-08)
+
+
+### Bug Fixes
+
+* make alias can be overridden properly ([#1648](https://github.com/vuepress/core/issues/1648)) ([13c8c59](https://github.com/vuepress/core/commit/13c8c59818e8cfbb25692ed1eaee04eef90fad1a))
+
+
+
 # [2.0.0-rc.22](https://github.com/vuepress/core/compare/v2.0.0-rc.21...v2.0.0-rc.22) (2025-04-28)